requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.6
grpcio==1.35.0
grpcio-tools==1.35.0
image==1.5.33
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.0
Pillow==8.1.0
protobuf==3.14.0
python-box==5.3.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
typing-extensions==3.7.4.3
wrapt==1.12.1
